// lib/drawings/site.generated.ts
export const LICENCE_URL = "/estate/LICENSE.txt";
// ...
export const POSTER: DrawingPoster = {
  eye: [568.423279, 568.423279, 337.691345], quat: [0.191342, 0.46194, 0.800103, 0.331414],
  vfovDeg: 42.184679, shift: [0, -0.123004], aspect: 1.333333, w: 1600, h: 1200,
};